+++ lib/recurly/fraud_info.php
@@ -7,8 +7,9 @@ class Recurly_FraudInfo
 {
   var $score;
   var $decision;
+  var $risk_rules_triggered;
 
   public function __toString() {
-    return "<Recurly_FraudInfo score=\"{$this->score}\" decision={$this->decision}>";
+    return "<Recurly_FraudInfo score=\"{$this->score}\" decision=\"{$this->decision}\">";
   }
 }