+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added `Recurly_AccountAcquisition` [#259](https://github.com/recurly/recurly-client-php/pull/259)
 * Added support for automated exports [#260](https://github.com/recurly/recurly-client-php/pull/260)
 * Added support for shipping addresses [#269](https://github.com/recurly/recurly-client-php/pull/269)
+* Adding support for `risk_rules_triggered` in `Recurly_FraudInfo` [#274](https://github.com/recurly/recurly-client-php/pull/274)
 
 ## Version 2.6.0 (August 9th, 2016)